requirements.txt
Quart==0.20.0
quart_cors==0.8.0
python-socketio==5.14.2
tortoise-orm==0.25.1
python-dotenv==1.2.1
uvicorn==0.38.0
cryptography==46.0.3
PyJWT==2.10.1
aioredis==2.0.1